store lat/lon in float
Yeah, the grid calculation is inaccurate.

$ clu CQ AA1BCD CQ LB2JK JO59 73 de K7IHZ DM43 call PM34ab
AA1BCD: country 179 'United States' cq 5 itu 8 continent 0 lat  37.60 lon -91.87
LB2JK @ JO59: country 197 'Norway' cq 14 itu 18 continent 4 lat  64.50 lon  21.00
K7IHZ @ DM43: country 179 'United States' cq 5 itu 8 continent 0 lat  38.50 lon -101.00
PM34ab: lat  39.56 lon 137.04

+++ src/dxcc.h
@@ -39,8 +39,8 @@ typedef struct
 	uchar cq; /* uchar max=255 */
 	uchar itu;
 	uchar continent;
-	int latitude;
-	int longitude;
+	float latitude;
+	float longitude;
 	short timezone;
 	const char *px;
 	const char *exceptions;
